DEV-6884 Removing int from lock apis
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@20692 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ toolkit-impl/src/main/java/com/terracotta/toolkit/collections/map/ServerMap.java
@@ -485,11 +485,11 @@ public class ServerMap<K, V> extends AbstractTCToolkitObject implements Internal
   }
 
   private static void beginLock(String lockID, final ToolkitLockTypeInternal type) {
-    ManagerUtil.beginLock(lockID, LockingUtils.translate(type).toInt());
+    ManagerUtil.beginLock(lockID, LockingUtils.translate(type));
   }
 
   private static void commitLock(String lockID, final ToolkitLockTypeInternal type) {
-    ManagerUtil.commitLock(lockID, LockingUtils.translate(type).toInt());
+    ManagerUtil.commitLock(lockID, LockingUtils.translate(type));
   }
 
   @Override
